// typings.d.ts
export interface Birth{
    first_name: string
    last_name: string
    department:string
    gender: string
    class: string
    moodle_id: number
    zodiac: string
}
export interface Birth2{
    first_name: string
    last_name: string
    department:string
    gender: string
    class: string
    moodle_id: number
    bday:string

}